security: fix HIGH gosec findings (17 → 0)
- graph/pubsub.go: subID generation used string(rune(n)) which
  produced mojibake past MaxRune and risked collisions on high
  values; switched to strconv.Itoa.
- transport/ws/hub.go: maxDropsBeforeClose is compared against an
  atomic int32 counter; changed the field type to int32 so the cast
  goes away.
- transport/gql/status.go: bounds-check HTTP status before storing
  as int32 (G115); out-of-range values become no-overrides.
- pubsub/rabbit/rabbit.go: clamp x-delivery-attempt to math.MaxInt32
  before narrowing.
- cmd/nexus/watcher.go: scanEmbedTargets now uses os.OpenRoot +
  rt.Open(rel) so a symlink swap during filepath.WalkDir can't
  escape the watched root (G122).
- app.go EnvAdminToken constant: #nosec G101 (env var name, not a
  credential value).
- client/vite_config.go (4 sites), cmd/nexus/{typefetcher,dev_pkg,
  deps}.go (6 sites): #nosec G703 with rationale — CLI helpers
  writing to operator-supplied paths, not a server vulnerability.

Cleanup: removed the activeDeploymentMatches stub left over from
the microsplit removal.

Verified: gosec ./... reports 0 HIGH findings; tests unchanged.

+++ transport/ws/hub.go
@@ -112,7 +112,7 @@ func (c *Connection) Send(message []byte) {
 	case c.send <- message:
 		c.drops.Store(0)
 	default:
-		if c.drops.Add(1) >= int32(c.hub.cfg.maxDropsBeforeClose) {
+		if c.drops.Add(1) >= c.hub.cfg.maxDropsBeforeClose {
 			c.hub.cfg.logf("closing slow client clientID=%s drops=%d", c.ClientID, c.drops.Load())
 			c.hub.unregister <- c
 		}
@@ -363,7 +363,7 @@ func (h *Hub) writeTo(c *Connection, data []byte) {
 	case c.send <- data:
 		c.drops.Store(0)
 	default:
-		if c.drops.Add(1) >= int32(h.cfg.maxDropsBeforeClose) {
+		if c.drops.Add(1) >= h.cfg.maxDropsBeforeClose {
 			h.unregister <- c
 		}
 	}
@@ -649,7 +649,7 @@ type hubConfig struct {
 	pongWait            time.Duration
 	pingPeriod          time.Duration
 	maxConnections      int
-	maxDropsBeforeClose int
+	maxDropsBeforeClose int32 // compared against atomic int32 drops counter
 	logf                func(format string, args ...any)
 }
